moved all files from "update IAM policy samples to use GAPIC library" PR to snippets module
