+++ iam/snippets/src/main/java/SetProjectPolicy.java
@@ -13,11 +13,9 @@
  * limitations under the License.
  */
 
-package iam.snippets;
-
-// [START iam_service_account_set_policy]
-import com.google.cloud.iam.admin.v1.IAMClient;
-import com.google.iam.admin.v1.ServiceAccountName;
+// [START iam_set_policy]
+import com.google.cloud.resourcemanager.v3.ProjectsClient;
+import com.google.iam.admin.v1.ProjectName;
 import com.google.iam.v1.Policy;
 import com.google.iam.v1.SetIamPolicyRequest;
 import com.google.protobuf.FieldMask;
@@ -25,36 +23,27 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 
-public class SetServiceAccountPolicy {
+public class SetProjectPolicy {
   public static void main(String[] args) throws IOException {
     // TODO(developer): Replace the variables before running the sample.
     // TODO: Replace with your project ID.
     String projectId = "your-project-id";
-    // TODO: Replace with your service account name.
-    String serviceAccount = "your-service-account";
     // TODO: Replace with your policy, GetPolicy.getPolicy(projectId, serviceAccount).
     Policy policy = Policy.newBuilder().build();
 
-    setServiceAccountPolicy(policy, projectId, serviceAccount);
+    setProjectPolicy(policy, projectId);
   }
 
-  // Sets a service account's policy.
-  public static Policy setServiceAccountPolicy(Policy policy, String projectId,
-                                               String serviceAccount) throws IOException {
-
-    // Construct the service account email.
-    // You can modify the ".iam.gserviceaccount.com" to match the service account name in which
-    // you want to delete the key.
-    // See, https://cloud.google.com/iam/docs/creating-managing-service-account-keys?hl=en#deleting
-
-    String accountEmail = String.format("%s@%s.iam.gserviceaccount.com", serviceAccount, projectId);
+  // Sets a project's policy.
+  public static Policy setProjectPolicy(Policy policy, String projectId)
+          throws IOException {
 
     // Initialize client that will be used to send requests.
     // This client only needs to be created once, and can be reused for multiple requests.
-    try (IAMClient iamClient = IAMClient.create()) {
+    try (ProjectsClient projectsClient = ProjectsClient.create()) {
       List<String> paths = Arrays.asList("bindings", "etag");
       SetIamPolicyRequest request = SetIamPolicyRequest.newBuilder()
-              .setResource(ServiceAccountName.of(projectId, accountEmail).toString())
+              .setResource(ProjectName.of(projectId).toString())
               .setPolicy(policy)
               // A FieldMask specifying which fields of the policy to modify. Only
               // the fields in the mask will be modified. If no mask is provided, the
@@ -63,8 +52,8 @@ public class SetServiceAccountPolicy {
               .setUpdateMask(FieldMask.newBuilder().addAllPaths(paths).build())
               .build();
 
-      return iamClient.setIamPolicy(request);
+      return projectsClient.setIamPolicy(request);
     }
   }
 }
-// [END iam_service_account_set_policy]
+// [END iam_set_policy]
