+++ iam/snippets/src/test/java/QuickstartTests.java
@@ -13,8 +13,6 @@
  * limitations under the License.
  */
 
-package iam.snippets;
-
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsCollectionContaining.hasItem;
